fixed build
git-svn-id: file:///home/maarten/qmmp-svn-raw/branches/qmmp-1.7@12518 90c681e8-e032-0410-971d-27865f9a5e38

+++ src/plugins/Ui/qsui/qsuivisualization.cpp
@@ -445,7 +445,7 @@ void QSUiAnalyzer::process(float *buffer, int width, int height)
 
     for(int i = 0; i < m_cols; i++)
     {
-        short y = 0;
+        int y = 0;
         int magnitude = 0;
 
         if(m_x_scale[i] == m_x_scale[i + 1])
